build: changes for E2E integration with cloud perch visualizations
